Moves auth module into `/modules`
