+++ src/modules/auth/auth.errors.ts
