+++ src/modules/auth/auth.model.ts
@@ -1,6 +1,6 @@
 import * as GQL from '@nestjs/graphql' // { Field, ObjectType, ID }
 
-import * as Utils from '../utils/Model'
+import * as Utils from '../../utils/Model'
 import { User } from '@entities/user/user.model'
 
 @GQL.ObjectType()
