+++ src/modules/auth/currentUser.ts
